+++ olap/src/aggregator/mod.rs
@@ -78,7 +78,7 @@ pub async fn run(
                                     Field::new(
                                         "value",
                                         DataType::Dictionary(
-                                            Box::new(DataType::Int8),
+                                            Box::new(DataType::Int32),
                                             Box::new(DataType::Binary),
                                         ),
                                         true,
@@ -308,7 +308,7 @@ mod tests {
                                                 Field::new(
                                                     "value",
                                                     DataType::Dictionary(
-                                                        Box::new(DataType::Int8),
+                                                        Box::new(DataType::Int32),
                                                         Box::new(DataType::Binary),
                                                     ),
                                                     true,
